Added a new class K8sServiceYamlParser; moved parsing of services from K8sYamlParser to the new class. Simplified service resource syntax check.
