+++ network-config-analyzer/K8sYamlParser.py
@@ -16,7 +16,7 @@ from GenericTreeScanner import TreeScannerFactory
 
 class K8sYamlParser(GenericYamlParser):
     """
-    A parser for k8s NetworkPolicy objects
+    A generic parser for k8s resources
     """
 
     def __init__(self, yaml_file_name=''):
@@ -205,132 +205,3 @@ class K8sYamlParser(GenericYamlParser):
             if re.fullmatch(r"[a-z0-9]([-a-z0-9]*[a-z0-9])?", port) is None:
                 self.syntax_error('port name should contain only lowercase alphanumeric characters or "-", '
                                   'and start and end with alphanumeric characters', port)
-
-    def parse_service(self, srv_object, peer_container):
-        """
-        Parses a service resource object and creates a K8sService object
-        :param dict srv_object: the service object to parse
-        :param PeerContainer peer_container: the peer container in which pods of the service are located
-        :return: K8sService object
-        """
-        if srv_object.get('kind') != 'Service':
-            return None  # Not a Service object
-        api_version = srv_object.get('apiVersion')
-        if api_version != 'v1':
-            return None  # apiVersion is not properly set
-        self.check_fields_validity(srv_object, 'Service', {'kind': [1, str], 'metadata': [1, dict],
-                                                           'spec': [1, dict], 'apiVersion': [1, str],
-                                                           'status': [0, dict]},
-                                   {'apiVersion': ['v1']})
-        metadata = srv_object['metadata']
-        allowed_metadata_keys = {'name': [1, str], 'annotations': [0, dict], 'creationTimestamp': [0, str],
-                                 'namespace': [0, str], 'resourceVersion': [0, str], 'uid': [0, str], 'labels': [0, dict]}
-        self.check_fields_validity(metadata, 'metadata', allowed_metadata_keys)
-        srv_name = metadata['name']
-        self.check_dns_subdomain_name(srv_name, metadata)
-        service = K8sService(srv_name)
-        if srv_object['spec'] is None:
-            self.warning('spec is missing or null in Service ' + srv_name)
-            return service
-
-        service_spec = srv_object['spec']
-        allowed_spec_keys = {'selector': [0, dict], 'type': [0, str], 'ports': [0, list],
-                             'clusterIP': [0, str], 'clusterIPs': [0, list], 'sessionAffinity': [0, str],
-                             'ipFamilies': [0, list], 'ipFamilyPolicy': [0, str], 'externalTrafficPolicy': [0, str]}
-        self.check_fields_validity(service_spec, 'spec', allowed_spec_keys,
-                                   {'type': ['ClusterIP', 'NodePort', 'LoadBalancer', 'ExternalName']})
-        service_type = service_spec.get('type', 'ClusterIP')
-        if service_type == 'ClusterIP':
-            service.set_type(K8sService.ServiceType.ClusterIP)
-        elif service_type == 'NodePort':
-            service.set_type(K8sService.ServiceType.NodePort)
-        elif service_type == 'LoadBalancer':
-            service.set_type(K8sService.ServiceType.LoadBalancer)
-        else:
-            service.set_type(K8sService.ServiceType.ExternalName)
-
-        selector = service_spec.get('selector')
-        if selector is not None:
-            for key, val in selector.items():
-                self.check_label_key_syntax(key, selector)
-                self.check_label_value_syntax(val, key, selector)
-                service.add_selector(key, val)
-                service.target_pods |= peer_container.get_peers_with_label(key, [val])
-
-        ports = service_spec.get('ports')
-        if ports is not None:
-            for port in ports:
-                self.check_fields_validity(port, 'Service port',
-                                           {'port': [1, int], 'protocol': [0, str], 'name': [0, str], 'targetPort': 0,
-                                            'nodePort': [0, int]},
-                                           {'protocol': ['TCP', 'UDP', 'SCTP']})
-                port_id = port.get('port', 0)
-                self.check_port_syntax(port_id, False, 'Service port')
-                target_port = port.get('targetPort')
-                if target_port is None:
-                    target_port = port_id
-                else:
-                    self.check_port_syntax(target_port, True, 'Service targetPort')
-                name = port.get('name', '')
-                self.check_dns_label_name(name, srv_object)
-                if not service.add_port(K8sService.ServicePort(port_id, target_port,
-                                                               port.get('protocol', 'TCP'), name)):
-                    self.warning(f'The port {name} is not unique in Service {srv_object}')
-                else:
-                    if isinstance(target_port, str):
-                        # check if all pods include this named port, and remove those that don't
-                        pods_to_remove = Peer.PeerSet()
-                        for pod in service.target_pods:
-                            pod_named_port = pod.named_ports.get(target_port)
-                            if not pod_named_port:
-                                self.warning(f'The named port {target_port} referenced in Service {srv_object} is not defined in the pod {pod}. Ignoring the pod')
-                                pods_to_remove |= pod
-                        service.target_pods -= pods_to_remove
-        return service
-
-    @staticmethod
-    def load_services_from_live_cluster(peer_container):
-        """
-        Loads and parses service resources from live cluster
-        :param PeerContainer peer_container: the peer container in which pods of the service are located
-        :return: The list of parsed services in K8sService format
-        """
-        PeerContainer.locate_kube_config_file()
-        yaml_file = CmdlineRunner.get_k8s_resources('service')
-        srv_resources = yaml.load(yaml_file, Loader=yaml.SafeLoader)
-        if not isinstance(srv_resources, dict):
-            return []
-        parser = K8sYamlParser('k8s')
-        res = []
-        for srv_code in srv_resources.get('items', []):
-            res.append(parser.parse_service(srv_code, peer_container))
-        return res
-
-    @staticmethod
-    def parse_service_resources(srv_resources_list, peer_container):
-        """
-        Parses the set of services in the container from one of the following resources:
-         - git path of yaml file or a directory with yamls
-         - local file (yaml or json) or a local directory containing yamls
-         - query of the cluster
-        :param list srv_resources_list: The service resource to be used.
-            If set to 'k8s', will query cluster using kubectl
-        :param PeerContainer peer_container: the peer container in which pods of the service are located
-        :return: The list of parsed services in K8sService format
-        """
-        for srv_resources in srv_resources_list:
-            # load from live cluster
-            if srv_resources == 'k8s':
-                return K8sYamlParser.load_services_from_live_cluster(peer_container)
-            else:
-                res = []
-                resource_scanner = TreeScannerFactory.get_scanner(srv_resources)
-                if resource_scanner is None:
-                    continue
-                yaml_files = resource_scanner.get_yamls()
-                for yaml_file in yaml_files:
-                    parser = K8sYamlParser(yaml_file)
-                    for srv_code in yaml_file.data:
-                        if isinstance(srv_code, dict) and srv_code.get('kind') in {'Service'}:
-                            res.append(parser.parse_service(srv_code, peer_container))
-                return res
